requirements.txt
# A2A AI Agent for Local Artisans - Requirements

# Core AI and Google services
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.40.0
google-cloud-storage>=2.10.0
vertexai>=1.40.0

# Streamlit web application
streamlit>=1.30.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0
pandas>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Data handling
dataclasses-json>=0.6.0

# Image processing
Pillow>=10.0.0

# Additional utilities
uuid>=1.30
logging>=0.4.9.6

# Development and testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0